ai-cap/chatbrain/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.6.1
PyYAML==6.0.1
rank-bm25==0.2.2
sentence-transformers==2.2.2
SQLModel==0.0.14
pytest==8.1.1
httpx==0.27.0
